Template fides over type
